requirements.txt
mcp>=1.2.0
Appium-Python-Client>=3.1.0
